Fold the MSAA tests into the GL viewer test module
Move both multi-sampling tests into test_viewer_get_frame.py rather than
carrying a test module of their own, and reuse that module's headless
ViewerGL helper, which already skips on the backend errors a machine
without a usable GL display raises.

Also fix the name of _setup_frame_buffer() in two comments.
